Support SignalFlow WebSocket message compression
Implement support for receiving compressed SignalFlow WebSocket
messages. Compression method used is GZIP, which is handled in
Javascript by the pako library.

If a message is compressed, it is transmitted over the WebSocket data
channel. Only the body of the message is compressed (after the 20-byte
binary message header). The body of the message, uncompressed, may be
either JSON (as indicated by the json flag), or binary.

For now, only data batch messages have their body encoded as binary, as
before.

The TextEncoding polyfill is used to handle decoding a string from the
decompressed Uint8Array buffer.