Remove winston logger
This commit removes winston logger as a dependency and replaces it with
a very simple console.log based implementation. The implementation
satisfies all current uses of the winston logger in this library.

This shaves off about 700 kilobyes un-gzipped from bundle size and
removes a few indirect dependencies which is very good for enviorments
like AWS Lambda.

This also makes it possible to build the library with esbuild which had
issues with one of it's dependencies (logform).