fix: Send remaining queue items if there are more than chunkSize
When the datapointsLength in a call to send() exceeded the chunkSize,
the excess items would remain in the queue until there was another call
with fewer than chunkSize datapoints.

If there were no future calls to send(), the items would remain in the
queue.

For now, only do this when the previous data is successfully sent to signalfx

Fixes #105